separate api logic
