+++ src/index.js
@@ -1,29 +1,15 @@
-/* eslint-disable no-unused-vars */
+/* eslint-disable no-unused-vars, import/no-cycle */
 import _, { get } from 'lodash';
 import 'bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './style.css';
+import { getScores, newScore } from './js/api';
 
-const requestUrl = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/0pCiglrVtP9Y26ppM44Z/scores/';
 const scoreBoard = document.querySelector('#scoresContainer');
 const refreshButton = document.querySelector('#refreshButton');
 const userName = document.querySelector('#userName');
 const userScore = document.querySelector('#userScore');
 const submitButton = document.querySelector('#submitScoreButton');
-
-async function newScore(user, score) {
-  const response = await fetch(requestUrl, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      user,
-      score,
-    }),
-  });
-  return response.json();
-}
 
 function populateBoard(obj) {
   const arr = obj.result;
@@ -38,20 +24,15 @@ function populateBoard(obj) {
   });
 }
 
-const request = new XMLHttpRequest();
-request.open('GET', requestUrl);
-request.responseType = 'json';
-request.send();
-request.onload = () => {
-  const scores = request.response;
-  populateBoard(scores);
-};
-
 refreshButton.addEventListener('click', () => {
-  populateBoard(request.response);
+  getScores();
 });
 
 submitButton.addEventListener('click', () => {
   newScore(userName.value, userScore.value);
-  populateBoard(request.response);
+  getScores();
 });
+
+getScores();
+
+export default populateBoard;
